feat: now milliseconds per tick can be set with thread safety

+++ src/main/scala/it/unibo/firesim/Main.scala
@@ -8,6 +8,6 @@ object Main:
   def main(args: Array[String]): Unit =
 
     val FPS = 30
-    val tickMs = (1000.0 / FPS).toLong
+    val tickMs = (1000.0 / FPS).toInt
     val controller = new SimController(new SimModel())
     controller.loop(tickMs)
